Update README.md
The documentation briefly describes the process of creating the mechanical parts of the robot and assembling the electronics, and discusses the process of training an object recognition model using deep learning from two different databases to enable the camera of a Raspberry Pi pocket computer to recognize different types of garbage. It describes the process by which the arm can estimate the distance of an object in space based on an ultrasonic and monocular depth estimation, and then, using a trigonometry-based reach, it can reach out to the appropriate distance for the garbage and pick it up. In addition, the paper explores the versatile applications of robotic arms such as this and the potential future developments associated with them, while discussing the motivation for the project.